requirements.txt
numpy~=1.22.3
pygame~=2.1.2
scipy~=1.8.0
pyopengl
